🚚 Move privacy component to privacy page
